+++ Xam.Zero/Forms/Xam.Zero.Dev/Xam.Zero.Dev/AppShell.xaml.cs
@@ -8,7 +8,5 @@ namespace Xam.Zero.Dev
         {
             this.InitializeComponent();
         }
-        
-        
     }
 }
