+++ Xam.Zero/Forms/Xam.Zero.Dev/Xam.Zero.Dev/Features/Home/HomeViewModel.cs
@@ -1,6 +1,7 @@
 using System.Threading.Tasks;
 using System.Windows.Input;
 using Xam.Zero.Dev.Features.Second;
+using Xam.Zero.ViewModels;
 using Xamarin.Forms;
 
 namespace Xam.Zero.Dev.Features.Home
