+++ Xam.Zero/Xam.Zero/Ioc/IContainer.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace Xam.Zero
+namespace Xam.Zero.Ioc
 {
     public interface IContainer
     {
